change suppr users for stripe

+++ app/Http/Controllers/BandController.php
@@ -125,9 +125,11 @@ class BandController extends Controller
             foreach($user->songs as $songtodelete){
             $this->deleteSong($songtodelete);
             }
-            Stripe::setApiKey(env("STRIPE_SECRET"));
-            $customer = \Stripe\Customer::retrieve($user->stripe_id);
-            $customer->delete();
+            if($user->stripe_id <> null){
+                Stripe::setApiKey(env("STRIPE_SECRET"));
+                $customer = \Stripe\Customer::retrieve($user->stripe_id);
+                $customer->delete();
+            }
             DB::table('subscriptions')->where('user_id',$user->id)->update(['stripe_status' => 'user_deleted', 'updated_at' => date('Y-m-d G:i:s')]);
             $user->delete();            
         }
